Split types from config module
